backend/requirements.txt
# Django and API dependencies
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Data processing and ML
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4

# Database (SQLite is included with Django)
# For production, you might want to add:
# psycopg2-binary==2.9.9  # For PostgreSQL
# mysqlclient==2.2.6      # For MySQL

# Additional utilities
python-dateutil==2.9.0
pytz==2024.2
